refactor(dispatch): extract safe modules (types/prompts/board/context-blocks)
Phase A of finishing the dispatch.ts split — the SAFE, non-fire-path modules. dispatch.ts is
the audit's most dangerous file (live exactly-once task dispatch), so the pure modules come
out first and the fire-path/mutex/singleton/cycle core stays untouched for a dedicated Phase B
with heavy rig dispatch-E2E.

Extracts 4 sibling modules (following the dispatch-failures.ts pattern):
- dispatch-types.ts (123) — the 14 shared types/interfaces; a leaf the other modules import from.
- dispatch-prompts.ts (210) — pure prompt builders (mcporterHelpers, sharedExecutionToolDocs,
  outputDisciplineSection, buildCorrectiveSection, buildDecompositionMessage, buildDispatchMessage)
  + IMAGE_MCPORTER_TIMEOUT_MS.
- dispatch-board.ts (102) — taskboard reads + policy + task-store wrappers (readDispatchColumns,
  isTaskDispatchEligible, findDispatchTaskSnapshot, addTaskLog/moveTaskToInProgress/tryAddTaskLog, ...).
- dispatch-context-blocks.ts (110) — sole owner of lessonBlockCache + buildDispatchLessonBlock/
  buildDispatchAssetBlock + __resetLessonBlockCache.

dispatch.ts: 2,096 → 1,652. It imports back what the core uses and re-exports the public surface
(isTaskDispatchEligible, buildDispatchMessage, buildDispatchAssetBlock, buildDispatchLessonBlock,
__resetLessonBlockCache, the Dispatch* types) so every @/core/dispatch consumer + the 3 tests that
mock.module the path are unchanged. No cycles (new modules import dispatch-types only). The
fire-path core (stateQueue mutex, fireDispatchTurn, dispatchTasks/dispatchSingleTask,
handleSessionDeath cycle, the dispatching/timer/inFlight/ladder singletons) is untouched.

Verified: typecheck + lint clean; dispatch/continuation/budget/task-service 96/0; full suite
5106/0; full binary build; boot smoke of the built binary (Dispatch started clean, no
module-load errors from the extracted modules).
